requirements.txt
Appium-Python-Client
pytest
requests